casting issues

+++ Models/Product.cs
@@ -10,7 +10,7 @@ namespace Team6.Models
         public int ProductId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
 
         public string ProductImage { get; set; }
